se agrego atributos a la clase enemigo (color, forma, velosidah)

+++ src/biz/randgames/hexattacks/entidades/Enemigo.java
@@ -5,7 +5,7 @@ package biz.randgames.hexattacks.entidades;
 public class Enemigo extends ObjetoEnMapa {
 	protected int hp;
 	protected int inmunidad;
-	protected int velocidad;
+	protected float velocidad;
